Setting

+++ src/pages/setting/Roles.jsx
@@ -1,22 +1,23 @@
-import React, { useRef } from "react";
+import React, { useRef, useCallback } from "react";
 import { useEffect } from "react";
 import { useTitle } from "../../hooks/useTitle";
-import "../../../node_modules/datatables.net-bs5/css/dataTables.bootstrap5.min.css";
-import "datatables.net-bs5";
-import $ from "jquery";
 import HeaderPage from "../../components/HeaderPage";
 import { useState } from "react";
 import Swal from "sweetalert2";
 import { SubmitOrCancelButton } from "../../components/SubmitOrCancelBtnForModal";
 import { useRole } from "../../hooks/roleStore";
 import { useNavigate } from "react-router-dom";
+import DataTableComponent from "../../components/DatatableComponent";
+import { Link } from "react-router-dom";
+import { handleCancel } from "../../util/handleCloseModal";
+import MainButton from "../../components/MainButton";
+import ModalComponent from "../../components/modal/ModalComponent";
 
 export const tableHead = [
-  { index: 0, colName: "ลำดับ" },
-  { index: 1, colName: "รหัสบทบาท" },
-  { index: 3, colName: "ชื่อบทบาท" },
-  { index: 4, colName: "คำอธิบาย" },
-  { index: 5, colName: "การจัดการ" },
+  { colName: "ลำดับ" },
+  { colName: "ชื่อบทบาท" },
+  { colName: "คำอธิบาย" },
+  { colName: "การจัดการ" },
 ];
 
 export default function Roles({ title }) {
@@ -38,14 +39,15 @@ export default function Roles({ title }) {
     deleteRole,
     updateRole,
     getRoleByIdData,
-    dataById
+    dataById,
   } = useRole();
-  const [editMode,setEditMode] = useState(false);
-  const [editRoleId,setEditRoleId] = useState(null);
+  const [editMode, setEditMode] = useState(false);
+  const [editRoleId, setEditRoleId] = useState(null);
+
   const navigate = useNavigate();
 
   const handleToPermissionPage = (roleId) => {
-    navigate(`/settings/rolepermission/:${roleId}`);
+    navigate(`/settings/rolepermission/${roleId}`);
   };
 
   const handleChangeInput = (e) => {
@@ -56,97 +58,58 @@ export default function Roles({ title }) {
     }));
   };
 
-  useEffect(() => {
-    const fetchDataTable = async () => {
-      try {
-        await getRoleData();
-      } catch (error) {
-        alert("ดึงข้อมูลไม่สำเร็จ", errorMessage);
-      }
-    };
-    fetchDataTable();
+  const fetchDataTable = useCallback(async () => {
+    try {
+      await getRoleData();
+    } catch (error) {
+      return;
+    }
   }, [getRoleData]);
 
-  //ตอนโหลตข้อมูลทั้งหมด
   useEffect(() => {
-    setEditMode(false)
-    if (data) {
-      GetDataTable();
-    }
-  }, [data]);
+    fetchDataTable();
+    setEditMode(false);
+  }, [fetchDataTable]);
 
-  useEffect(()=>{
-   if(dataById){
+  useEffect(() => {
+    if (dataById) {
       setInput((prevData) => ({
         ...prevData,
-        roleName : dataById.roleName  ?? "",
-        description : dataById.description ?? ""
+        roleName: dataById.roleName ?? "",
+        description: dataById.description ?? "",
       }));
     }
-  },[dataById])
+  }, [dataById]);
 
   useEffect(() => {
     if (Object.keys(error).length === 0 && isSubmit) {
     }
   }, [error, isSubmit]);
 
-  const GetDataTable = () => {
-    $(tableRef.current).DataTable({
-      data: data,
-      destroy: true,
-      responsive: true,
-      paging: true,
-      searching: true,
-      autoWidth: true,
-      language: {
-        decimal: "",
-        emptyTable: "ไม่มีข้อมูลในตาราง",
-        info: "แสดง _START_ ถึง _END_ จาก _TOTAL_ รายการ",
-        infoEmpty: "แสดง 0 ถึง 0 จาก 0 รายการ",
-        infoFiltered: "(กรองจาก _MAX_ รายการทั้งหมด)",
-        infoPostFix: "",
-        thousands: ",",
-        lengthMenu: "แสดง _MENU_ รายการ",
-        loadingRecords: "กำลังโหลด...",
-        processing: "กำลังประมวลผล...",
-        search: "ค้นหา:",
-        zeroRecords: "ไม่พบข้อมูลที่ตรงกัน",
+  const columnData = [
+    {
+      data: null,
+      render: function (data, type, row, meta) {
+        return meta.row + 1;
       },
-      columnDefs: [
-        { width: "70px", targets: 0 },
-        { width: "70px", targets: 1 },
-        { width: "200px", targets: 2 },
-        { width: "230px", targets: 3 },
-        { width: "120px", targets: 4 },
-      ],
-      columns: [
-        {
-          data: null,
-          render: function (data, type, row, meta) {
-            return meta.row + 1;
-          },
-        },
-        {
-          title: "รหัสบทบาท",
-          data: "roleId",
-          orderable: true,
-        },
-        {
-          title: "บทบาท",
-          data: "roleName",
-          orderable: true,
-        },
+    },
 
-        {
-          title: "คำอธิบาย",
-          data: "description",
-          orderable: true,
-        },
-        {
-          data: null,
-          title: "การจัดการ",
-          render: function (data, type, row) {
-            return `      
+    {
+      title: "บทบาท",
+      data: "roleName",
+      orderable: true,
+    },
+
+    {
+      title: "คำอธิบาย",
+      data: "description",
+      orderable: true,
+    },
+    {
+      data: null,
+      title: "การจัดการ",
+      render: function (data, type, row) {
+        return `      
          <div className="d-flex align-items-center justify-content-center">
             <div class="dropdown d-lg-none">
               <button class="btn btn-outline-light" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
@@ -154,17 +117,25 @@ export default function Roles({ title }) {
               </button>
               <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
                 <li>
-                <a class="dropdown-item text-dark btn-edit" data-role-id="${row.roleId}">
+                <a class="dropdown-item text-dark btn-edit" 
+                   data-id="${row.roleId}"
+                   data-action="edit" 
+                   >
                   <i class="bi bi-pen-fill me-2"></i> แก้ไขข้อมูล
                 </a>
               </li>
               <li>
-                <a class="dropdown-item text-dark btn-permission" data-role-id="${row.roleId}">
+                <a class="dropdown-item text-dark btn-permission" 
+                   data-id="${row.roleId}"
+                   data-action="permission"
+                   >
                   <i class="bi bi-person-fill-lock me-2"></i>สิทธิเข้าใช้งาน
                 </a>
               </li>
               <li>
-                <a class="dropdown-item text-dark btn-delete" data-role-id="${row.roleId}">
+                <a class="dropdown-item text-dark btn-delete"
+                   data-id="${row.roleId}" 
+                   data-action="delete">
                   <i class="bi bi-trash-fill me-2"></i> ลบข้อมูล
                 </a>
               </li>
@@ -176,52 +147,50 @@ export default function Roles({ title }) {
               
               class="btn btn-warning me-2 btn-edit"
               title="แก้ไข"
-              data-role-id="${row.roleId}"
+              data-id="${row.roleId}"
+              data-action="edit" 
             >
               <i class="bi bi-pen-fill"></i>
             </a>
             <a
               class="btn btn-info me-2 btn-permission"
               title="สิทธิเข้าใช้งาน"
-              data-role-id="${row.roleId}"
+              data-id="${row.roleId}"
+               data-action="permission"
             >
               <i class="bi bi-person-fill-lock"></i>
             </a>
             <a
               class="btn btn-danger btn-delete"
               title="ลบ"
-              data-role-id="${row.roleId}"
+              data-id="${row.roleId}"
+              data-action="delete"
             >
               <i class="bi bi-trash-fill"></i>
             </a>
           </div>
         </div>
        `;
-          },
-        },
-      ],
-      dom:
-        window.innerWidth <= 570
-          ? '<"top"lf>rt<"bottom"ip><"clear">'
-          : '<"top"lf>rt<"bottom"ip><"clear">',
-    });
+      },
+    },
+  ];
+
+  const columnDefs = [
+    { width: "70px", targets: 0, className: "text-center" },
+    { width: "100px", targets: 1 },
+    { width: "150px", targets: 2, className: "fs-6" },
+    { width: "90px", targets: 3 },
+  ];
+
+  const handleAction = (action, id) => {
+    if (action === "edit") {
+      handleEdit(id);
+    } else if (action === "delete") {
+      handleDelete(id);
+    } else if (action === "permission") {
+      handleToPermissionPage(id);
+    }
   };
-
-  //  เข้าสู่ ปุ่ม action
-  $(tableRef.current).on("click", ".btn-permission", function () {
-    const id = $(this).data("roleId");
-    handleToPermissionPage(id);
-  });
-
-  $(tableRef.current).on("click", ".btn-delete", function () {
-    const id = $(this).data("roleId");
-    handleDelete(id);
-  });
-
-  $(tableRef.current).on("click", ".btn-edit", function () {
-    const id = $(this).data('roleId');
-    handleEdit(id);
-  });
 
   const validateForm = () => {
     let errors = {};
@@ -231,7 +200,7 @@ export default function Roles({ title }) {
     return errors;
   };
 
-  const handleSubmit = async (e) => {
+  const handleSubmit = async (e, modalId) => {
     e.preventDefault();
     const reqData = {
       roleName: input.roleName,
@@ -242,41 +211,78 @@ export default function Roles({ title }) {
     setError(errorList);
     //api post
     if (Object.keys(errorList).length === 0) {
-      const response = editMode? await updateRole(reqData,editRoleId) : await createRole(reqData);
-      if (response.success) {
-        setIsSubmit(true);
-        Swal.fire({
-          title: "บันทึกข้อมูลสำเร็จ",
-          icon: "success",
-          draggable: true,
-          buttonsStyling: "w-100",
-        });
+      const swalWithBootstrapButtons = Swal.mixin({
+        customClass: {
+          confirmButton: "btn btn-success custom-width-btn-alert",
+          cancelButton: "btn btn-danger custom-width-btn-alert",
+        },
+        buttonsStyling: "w-100",
+      });
+      swalWithBootstrapButtons
+        .fire({
+          title: "คุณต้องการบันทึกรายการใช่หรือไม่",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "ยื่นยันการบันทึกรายการ",
+          cancelButtonText: "ยกเลิกการบันทึกรายการ",
+          reverseButtons: true,
+        })
+        .then(async (result) => {
+          if (result.isConfirmed) {
+            const { errorMessage, success } = editMode
+              ? await updateRole(reqData, editRoleId)
+              : await createRole(reqData);
+            if (success) {
+              swalWithBootstrapButtons.fire({
+                title: "บึนทึกรายการสำเร็จ!",
+                icon: "success",
+              });
 
-        const currentModal = document.getElementById("notModal");
-        const modalInstance = bootstrap.Modal.getInstance(currentModal);
-        modalInstance.hide();
-        ClearInput();
-        await getRoleData();
-      } else {
-        Swal.fire({
-          title: "บันทึกข้อมูลไม่สำเร็จ",
-          icon: "error",
+              const currentModal = document.getElementById(modalId);
+              const modalInstance = bootstrap.Modal.getInstance(currentModal);
+              modalInstance.hide();
+
+              ClearInput();
+              await getRoleData();
+            } else {
+              Swal.fire({
+                title: "บันทึกข้อมูลไม่สำเร็จ",
+                text: errorMessage,
+                icon: "error",
+              });
+            }
+          } else if (result.dismiss === Swal.DismissReason.cancel) {
+            swalWithBootstrapButtons.fire({
+              title: "ยกเลิก",
+              text: "คุณทำการยกเลิกรายการเรียบร้อยแล้ว",
+              icon: "error",
+            });
+          }
         });
-      }
     }
   };
 
-  const handleEdit =async (roleId) => {
-    await getRoleByIdData(roleId); //ผูก api เรียกใช้ข้อมูล
-    setEditRoleId(roleId)
+  const handleOpenModal = () => {
+    ClearInput();
+    setEditMode(false);
+    const currentModal = document.getElementById("notModal");
+    if (currentModal) {
+      const modal = bootstrap.Modal.getOrCreateInstance(currentModal);
+      modal.show();
+    }
+  };
+
+  const handleEdit = async (roleId) => {
+    ClearInput();
+    await getRoleByIdData(roleId);
+    setEditRoleId(roleId);
 
     const currentModal = document.getElementById("notModal");
     if (currentModal) {
       //เป็นการสร้างใหม่ ก่อนการเรียกใช้
       const modal = bootstrap.Modal.getOrCreateInstance(currentModal);
       modal.show();
-      setEditMode(true)
-      
+      setEditMode(true);
     }
   };
 
@@ -334,11 +340,11 @@ export default function Roles({ title }) {
   };
 
   return (
-    <div className="container py-4 min-vh-90 d-flex flex-column">
+    <div className="container-fluid py-4 min-vh-90 d-flex flex-column">
       <nav aria-label="breadcrumb">
         <ol className="breadcrumb">
           <li className="breadcrumb-item">
-            <a href="/settings">ตั้งค่า</a>
+            <Link to="/settings">ตั้งค่า</Link>
           </li>
           <li className="breadcrumb-item active" aria-current="page">
             {title}
@@ -348,46 +354,89 @@ export default function Roles({ title }) {
       <HeaderPage pageName={title} />
       <div className="container">
         {/* ปุ่มเพิ่ม */}
-        <div className="add-btn">
-          <a
-            className="power py-2"
-            style={{ maxWidth: "200px" }}
-            data-bs-toggle="modal"
-            data-bs-target="#notModal"
-          >
-            <span>
-              <i class="bi bi-plus-circle fs-4"></i>
-            </span>{" "}
-            <span className="label">{addBtnName}</span>
-          </a>
-        </div>
-        <div className="mt-4">
-          <table
-            ref={tableRef}
-            className="table table-striped"
-            style={{ width: "100%" }}
-          >
-            <thead>
-              <tr>
-                {tableHead.map((row) => (
-                  <th
-                    key={row.index}
-                    style={{
-                      background: "#ffe8da",
-                      fontWeight: "600",
-                      padding: "12px 8px",
-                    }}
-                  >
-                    {row.colName}
-                  </th>
-                ))}
-              </tr>
-            </thead>
-          </table>
-        </div>
+        <MainButton
+          btnName={addBtnName}
+          icon={"bi bi-plus-circle"}
+          onClick={handleOpenModal}
+        />
+        {/* table */}
+
+        <DataTableComponent
+          column={columnData}
+          data={data}
+          onAction={handleAction}
+          tableHead={tableHead}
+          tableRef={tableRef}
+          columnDefs={columnDefs}
+        />
 
         {/* modal */}
-        <div
+
+        <ModalComponent
+          icon="bi bi-plus-circle"
+          modalId="notModal"
+          modalSize="modal-md"
+          title={title}
+        >
+          <div className="employee-content p-4">
+            <div
+              style={{
+                display: "flex",
+                flexDirection: "column",
+                alignItems: "center",
+                width: "100%",
+              }}
+            >
+              <form>
+                <div>
+                  <div className="mb-3">
+                    <label className="form-label">
+                      ชื่อบทบาท
+                      <span style={{ color: "red" }}>*</span>
+                    </label>
+                    <input
+                      name="roleName"
+                      type="text"
+                      className={`form-control ${
+                        error.roleName ? "border border-danger" : ""
+                      }`}
+                      placeholder="กรอกชื่อบทบาท"
+                      value={input.roleName}
+                      onChange={handleChangeInput}
+                    />
+                    {error.roleName ? (
+                      <p className="text-danger">{error.roleName}</p>
+                    ) : null}
+                  </div>
+                  <div className="mb-3">
+                    <label className="form-label">
+                      คำอธิบายบทบาท
+                      <span style={{ color: "red" }}>*</span>
+                    </label>
+                    <textarea
+                      className="form-control"
+                      name="description"
+                      placeholder="กรอกคำอธิบาย"
+                      rows="3"
+                      cols="40"
+                      maxlength="100"
+                      style={{ resize: "none", overflow: "hidden" }}
+                      value={input.description}
+                      onChange={handleChangeInput}
+                    ></textarea>
+                  </div>
+                </div>
+              </form>
+            </div>
+          </div>
+          <SubmitOrCancelButton
+            handleSubmit={(e) => handleSubmit(e, "notModal")}
+            handleCancel={() => handleCancel("notModal")}
+            isLoading={isLoading}
+          />
+          {isLoading && <span className="loader"></span>}
+        </ModalComponent>
+        {/* <div
           className="modal fade"
           id="notModal"
           tabIndex="-1"
@@ -407,10 +456,10 @@ export default function Roles({ title }) {
                   className="btn-close"
                   data-bs-dismiss="modal"
                   aria-label="Close"
-                  onClick={ClearInput}
+                  // onClick={() => handleCancel("notModal")}
                 ></button>
               </div>
-              <div class="modal-body">
+              <div className="modal-body">
                 <div className="employee-content p-4">
                   <div
                     style={{
@@ -421,10 +470,9 @@ export default function Roles({ title }) {
                     }}
                   >
                     <form>
-                      {/* ข้อมูลทั่วไป */}
                       <div>
                         <div className="mb-3">
-                          <label class="form-label">
+                          <label className="form-label">
                             ชื่อบทบาท
                             <span style={{ color: "red" }}>*</span>
                           </label>
@@ -458,7 +506,6 @@ export default function Roles({ title }) {
                             value={input.description}
                             onChange={handleChangeInput}
                           ></textarea>
-                         
                         </div>
                       </div>
                     </form>
@@ -466,13 +513,14 @@ export default function Roles({ title }) {
                 </div>
               </div>
               <SubmitOrCancelButton
-                handleSubmit={handleSubmit}
-                handleCancel={ClearInput}
+                handleSubmit={(e) => handleSubmit(e, "notModal")}
+                handleCancel={() => handleCancel("notModal")}
+                isLoading={isLoading}
               />
               {isLoading && <span className="loader"></span>}
             </div>
           </div>
-        </div>
+        </div> */}
       </div>
     </div>
   );
